Refactor `AudienceTiles`.
